requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
apscheduler
python-nmap
paho-mqtt
pyyaml
